Use workflows from the git submodule instead of github
ok heres the deal:
All the references to "@main" means that if the user wants to pin KSPBuildTools to a specific version or commit (perfectly reasonable if theres ever breaking changes), then any sub-action will still call from the main branch. Alternatively we can just use the file directly from the git submodule, which is needed anyways for KSPCommon.props. That way version pinning is done entirely through the git submodule to a specific commit