Change of strategy
Instead of a couple big uber-configurable actions, have a handful of small ones. its still only 6 steps total in the template workflow.

* ckan-setup: install CKAN, setup an instance, install any mods listed
* install-dependencies: run the CKANInstall target in msbuild
* compile: Same as before, but no longer calls the previous 2 steps

all or none of these can be used by the user depending on what their mod is and how it is set-up

I think this will be a lot more modular and easy to use?